requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
Flask_Limiter==3.3.0
geojson==3.0.1
psycopg2==2.9.5
python-dotenv==1.0.0
Shapely==2.0.1
waitress==2.1.2
